#1363 Add ProxyDatasourceType to split atomikos and tomcat-dbcp2.
